refactor: move RunnerOutput into a new mod
