+++ src/lib.rs
@@ -8,4 +8,5 @@ pub mod request;
 pub mod response;
 pub mod router;
 pub mod runner_input;
+pub mod runner_output;
 pub mod server;
